pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "Windx"
version = "1.0.0"
description = "Automated product configurator for custom manufacturing - windows, doors, furniture, and more"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "alaamer12", email = "ahmedmuhmmed239@gmail.com" }
]
keywords = [
    "product-configurator",
    "manufacturing",
    "custom-products",
    "pricing-engine",
    "quote-generation",
    "order-management",
    "fastapi",
    "postgresql",
    "ltree",
    "hierarchical-data",
    "e-commerce",
    "b2b",
    "windows",
    "doors",
    "furniture"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Manufacturing",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Office/Business",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Operating System :: OS Independent",
]

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "asyncpg",
    "pydantic",
    "pydantic-settings",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "alembic",
    "fastapi-cache2[redis]",
    "fastapi-limiter",
    "fastapi-pagination",
    "redis",
    "jinja2",
    "bcrypt==4.3.0", # Must be version 4
    "python-multipart",
    "matplotlib",
    "networkx",
    # "pygraphviz",
    "pandas>=2.3.3",
    "rich",
    "platformdirs",
    "requests",
    "email-validator", # Required by pydantic even if it is not used explicitly
    "casbin",
    "casbin-sqlalchemy-adapter",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-mock",
    "pytest-xdist",           # Parallel test execution
    "pytest-sugar",           # Better test output
    "pytest-clarity",         # Better assertion diffs
    "pytest-icdiff",          # Improved diff output
    "pytest-html",            # HTML test reports
    "pytest-json-report",     # JSON test reports
    "pytest-emoji",           # Emoji test output
    "pytest-instafail",       # Show failures instantly
    "pytest-picked",          # Run only changed tests
    "ruff",
    "httpx",
    "playwright",
    "pytest-playwright",
    "psutil",
    "hypothesis",
    "aiohttp"
]

[project.urls]
Homepage = "https://github.com/alaamer12/Windx"
Documentation = "https://github.com/alaamer12/Windx#readme"
Repository = "https://github.com/alaamer12/Windx"
Issues = "https://github.com/alaamer12/Windx/issues"
Changelog = "https://github.com/alaamer12/Windx/releases"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "UP045", # Old types pattern
    "W293",
    "B904",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",                           # Verbose output
    "-l",                           # Show local variables in tracebacks
    "-ra",                          # Show all test results summary
    "--strict-markers",             # Strict marker checking
    "-W", "default",                # Show warnings
    "--tb=short",                   # Short traceback format
    "--cov-report=term-missing",    # Terminal coverage with missing lines
    # "--maxfail=5",                # Commented out - let CI control this
    "--instafail",                  # Show failures instantly
    # Optional plugins - only enabled if available:
    # "--emoji",                    # Use emoji in output (pytest-emoji)
    # "--sugar",                    # Better test output formatting (pytest-sugar)
    # "--clarity",                  # Better assertion diffs (pytest-clarity)
]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (slower, full stack)",
    "e2e: End-to-end tests (browser-based, requires running app)",
    "slow: Slow tests",
    "auth: Authentication tests",
    "users: User management tests",
    "services: Service layer tests",
    "repositories: Repository layer tests",
    "redis: Tests requiring Redis (cache and rate limiting)",
    "ci_cd_issue: Tests that have known issues in CI/CD environment but pass locally",
]
log_cli = false
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# Enhanced error reporting
junit_family = "xunit2"
junit_logging = "all"
junit_log_passing_tests = false
